requirements.txt
python-telegram-bot>=20.0,<21.0
aiosqlite>=0.19.0
APScheduler>=3.10.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

